ohai: Treat struct as fixed length instead of zero terminated string
While the result gets written to the existing memory of the ecmd
variable, unpack it again to avoid violating any assumptions that
existing memory doesn't change. (Testing didn't show anything making
this assumption, but just to be safe.) When unpack("p") is used on a
struct it might read too many or not enough bytes depending on when it
finds a zero. There is no size for pack("P") needed, only on unpack.