network: Raise exception if conduit does not exist in mapping
With the ability to override a conduit for a network for a specific
node, there's no early validation that a conduit actually exists in the
conduit mapping, and this crashes in an exception that is impossible to
understand.

Let's add a check and raise an explicit exception in that case.