network: Avoid trying to plug the same interface to two ovs bridges (bsc#1120657)
Before attaching a base interface to an ovs bridge check whether it is
already connected to a different bridge.

If that other bridge is still needed by the current network
configuration the recipe will error out and report a conflict in the
current network configuration.

If the other bridge is no longer needed, the base interface is unplugged
from the bridge and the bridge is shutdown and taken away from the
control of wicked.